Moved database models to db package Created rest package with corresponding DTO (data transfer objects)
